Split optimize report generation from visualization
