+++ freqtrade/optimize/optimize_reports/__init__.py
@@ -4,7 +4,13 @@ from freqtrade.optimize.optimize_reports.optimize_reports import (
     generate_edge_table, generate_exit_reason_stats, generate_pair_metrics,
     generate_periodic_breakdown_stats, generate_rejected_signals, generate_strategy_comparison,
     generate_strategy_stats, generate_tag_metrics, generate_trade_signal_candles,
-    generate_trading_stats, generate_wins_draws_losses, show_backtest_result, show_backtest_results,
-    show_sorted_pairlist, store_backtest_analysis_results, store_backtest_stats,
-    text_table_add_metrics, text_table_bt_results, text_table_exit_reason,
-    text_table_periodic_breakdown, text_table_strategy, text_table_tags)
+    generate_trading_stats, generate_wins_draws_losses, store_backtest_analysis_results,
+    store_backtest_stats)
+from freqtrade.optimize.optimize_reports.visualization import (show_backtest_result,
+                                                               show_backtest_results,
+                                                               show_sorted_pairlist,
+                                                               text_table_add_metrics,
+                                                               text_table_bt_results,
+                                                               text_table_exit_reason,
+                                                               text_table_periodic_breakdown,
+                                                               text_table_strategy, text_table_tags)
